+++ include/onewire.h
@@ -10,7 +10,7 @@
 void own_scanbus(uint8_t portnum);
 int16_t own_readtemp(uint8_t portnum, int16_t * temp, int16_t * t_frac);
 /* 0 = success, -1 = error */
-int8_t get_persistent_mac(uint8_t portnum, uint8_t * mac);
-int8_t set_persistent_mac(uint8_t portnum, uint8_t * mac);
+int get_persistent_mac(uint8_t portnum, uint8_t * mac);
+int set_persistent_mac(uint8_t portnum, uint8_t * mac);
 
 #endif
